+++ src/fake-schema/field/directives/fake/fakers/index.ts
@@ -2,6 +2,6 @@ export { Fakers, createTypeFakers } from "./Fakers";
 export { Random } from "./Random";
 export { TypeFakers } from "./TypeFakers";
 import * as resolve from "./resolve";
-import * as maps from "./maps";
+import * as maps from "../maps";
 import * as functions from "./functions";
 export { resolve, maps, functions };
