+++ src/fake-schema/field/directives/fake/fakers/resolve/example/ExampleResolver.ts
@@ -1,4 +1,4 @@
-import { examples as exampleMaps } from "../fake/fakers/maps";
+import { examples as exampleMaps } from "../../../maps";
 import {
   createKeyMatcher,
   resolveFromFieldMap,
