+++ src/fake-schema/field/directives/fake/maps/index.ts
@@ -1,4 +1,3 @@
 import { examples } from "./examples";
 import { fakes } from "./fakes";
-export const maps = { examples, fakes };
 export { examples, fakes };
